+++ internal/domain/event/tag_deleted_event.go
@@ -1,0 +1,31 @@
+package event
+
+import (
+	"github.com/kipitix/growscada/internal/domain/tag"
+)
+
+type TagDeletedEvent interface {
+	TagEvent
+}
+
+type tagDeletedEventImpl struct {
+	tagEventImpl
+}
+
+var _ TagDeletedEvent = (*tagDeletedEventImpl)(nil)
+
+func NewTagDeletedEvent(anID tag.TagID, opts ...EventOption) TagDeletedEvent {
+	ev := &tagDeletedEventImpl{
+		tagEventImpl: tagEventImpl{
+			tagID: anID,
+			eventImpl: eventImpl{
+				eventType: EventTypeTagDeleted,
+				timestamp: NewEventTimestamp(),
+			},
+		},
+	}
+
+	ev.applyOptions(opts...)
+
+	return ev
+}