+++ internal/domain/event/tag_updated_event.go
@@ -1,0 +1,31 @@
+package event
+
+import (
+	"github.com/kipitix/growscada/internal/domain/tag"
+)
+
+type TagUpdatedEvent interface {
+	TagEvent
+}
+
+type tagUpdatedEventImpl struct {
+	tagEventImpl
+}
+
+var _ TagUpdatedEvent = (*tagUpdatedEventImpl)(nil)
+
+func NewTagUpdatedEvent(aTagID tag.TagID, opts ...EventOption) TagUpdatedEvent {
+	ev := &tagUpdatedEventImpl{
+		tagEventImpl: tagEventImpl{
+			tagID: aTagID,
+			eventImpl: eventImpl{
+				eventType: EventTypeTagUpdated,
+				timestamp: NewEventTimestamp(),
+			},
+		},
+	}
+
+	ev.applyOptions(opts...)
+
+	return ev
+}