+++ internal/domain/tag/tag_value.go
@@ -1,8 +1,11 @@
 package tag
 
+import "fmt"
+
 // TagValue represents the value of a tag.
 // Value Object
 type TagValue interface {
 	Value() any
-	String() string
+
+	fmt.Stringer
 }